Redownload the data and use all of the countries
The URL changed to https://berkeleyearth.org and some other things
changed about the website. Redownload all the data and store it in the
repository this time. Process but keep the monthly average (which is
noisy but it's more complete and can be used to teach filtering), the
uncertainty, and the year and month. Add a header to each file
explaining provenance. Add a CC-BY-NC license to the repo to be in line
with Berkeley Earth. Add download info to the data README.